requirements.txt
fastapi
uvicorn
python-dotenv
numpy
opencv-python
onnxruntime
ultralytics
scikit-learn
supabase
python-jose
requests
